BE/requirements.txt
flask==1.1.2
numpy==1.19.1
pillow==7.2.0
scikit-image==0.17.2
torch==1.6.0
torchvision==0.7.0
waitress==1.4.4
tqdm==4.48.2
requests==2.24.0
scipy==1.5.2
pymatting==1.0.6
filetype==1.0.8
rembg==1.0.14
flask_sqlalchemy>=2.5.0
flask_restx==0.5.1
flask_login==0.5.0